requirement.txt
torch
torchvision
flask==2.0.1
werkzeug
pillow
numpy